jci_interview_problem/message_parser.py
"""Parses messages with a provided data source and search criteria."""

import concurrent.futures
import logging


def parse(message_file, segment='', field_name='', limit=None):
    """Parses an open file handle for messages matching the provided search criteria.

    Args:
        message_file (File): A file object containint 0 or more messages.
        segment (Optional(str)): A segment to search for. If empty, results are not filtered by segment.
        field_name (Optional(str)): A field name to search for. If empty, results are not filtered by field name.
        limit (Optional(int)): The number of messages to return.

    Returns:
        List[str]: The matching messages.
    """
    return None
